Move the tag code to a concern
This is to allow it to be used with multiple models. It does require a
`TAG_TYPE` constant to be set on the base model. This is used to set the
prefix used in the provider tags
