+++ lib/cloudware/cli.rb
@@ -27,6 +27,10 @@ require 'exceptions'
 require 'command'
 
 require 'require_all'
+
+require_all 'lib/cloudware/models/concerns/**/*.rb'
+require_all 'lib/cloudware/models/**/*.rb'
+
 require_all 'lib/cloudware/commands/concerns/**/*.rb'
 require_all 'lib/cloudware/commands/**/*.rb'
