+++ lib/cloudware/models/concerns/tags.rb
@@ -1,0 +1,57 @@
+# frozen_string_literal: true
+
+module Cloudware
+  module Models
+    module Concerns
+      module Tags
+        # NOTE: The TAG_TYPE must be set as a constant on the base model
+
+        def self.included(base)
+          base.extend(ClassMethods)
+        end
+
+        module ClassMethods
+          def prefix
+            "cloudware#{self::TAG_TYPE}"
+          end
+
+          def flag
+            'TAG'
+          end
+
+          def build_from_deployment(deployment)
+            (deployment.results || {})
+                       .keys
+                       .map { |k| self.name_from_tag(k) }
+                       .uniq
+                       .reject { |n| n.nil? }
+                       .map do |name|
+              self.new(name: name, deployment: deployment)
+            end
+          end
+
+          def name_from_tag(tag)
+            regex = /(?<=\A#{prefix}).*(?=#{flag}.*\Z)/
+            regex.match(tag.to_s)&.to_a&.first
+          end
+
+          def tag_generator(name, tag)
+            :"#{prefix}#{name}#{flag}#{tag}"
+          end
+        end
+
+        def tags
+          (deployment.results || {}).each_with_object({}) do |(key, value), memo|
+            next unless (tag = extract_tag(key))
+            memo[tag] = value
+          end
+        end
+
+        def extract_tag(key)
+          regex = /(?<=\A#{self.class.tag_generator(Regexp.escape(name), '')}).*/
+          regex.match(key.to_s)&.to_a&.first&.to_sym
+        end
+      end
+    end
+  end
+end
